requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# PostGIS
geoalchemy2==0.14.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
python-csv==0.0.2

# Visualizations
plotly==5.18.0
folium==0.15.1
kaleido==0.2.1

# Export
reportlab==4.0.8

# HTMX and Frontend
jinja2==3.1.3

# Redis Cache
redis==5.0.1
aiocache==0.12.2

# HTTP Client for Kobo
httpx==0.26.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
ephem==4.1.5
